Convert EVERYTHING to postgres

+++ setup.sql
@@ -1,5 +1,4 @@
-CREATE TABLE
-IF NOT EXISTS tags
+CREATE TABLE IF NOT EXISTS tags
 (
     tag_id TEXT,
     content TEXT,
@@ -8,15 +7,13 @@ IF NOT EXISTS tags
     views INTEGER
 );
 
-CREATE TABLE
-IF NOT EXISTS tag_relations
+CREATE TABLE IF NOT EXISTS tag_relations
 (
     tag_id TEXT,
     alias TEXT
 );
 
-CREATE TABLE
-IF NOT EXISTS blacklist
+CREATE TABLE IF NOT EXISTS blacklist
 (
     user_id BIGINT,
     reason TEXT,
@@ -26,18 +23,16 @@ IF NOT EXISTS blacklist
     expires BIGINT
 );
 
-CREATE TABLE
-IF NOT EXISTS flag_quizz
+CREATE TABLE IF NOT EXISTS flag_quizz
 (
     user_id BIGINT,
     tries INTEGER,
     correct INTEGER,
     completed INTEGER,
-    guild_id INTEGER
+    guild_id BIGINT
 );
 
-CREATE TABLE
-IF NOT EXISTS trivia
+CREATE TABLE IF NOT EXISTS trivia
 (
     id BIGINT,
     correct INTEGER,
@@ -46,8 +41,7 @@ IF NOT EXISTS trivia
     longest_streak INTEGER
 );
 
-CREATE TABLE
-IF NOT EXISTS reaction_roles
+CREATE TABLE IF NOT EXISTS reaction_roles
 (
     message_id BIGINT,
     role_id BIGINT,
@@ -55,8 +49,7 @@ IF NOT EXISTS reaction_roles
     roles_given INTEGER DEFAULT 0
 );
 
-CREATE TABLE
-IF NOT EXISTS warnings
+CREATE TABLE IF NOT EXISTS warnings
 (
     warning_id SERIAL PRIMARY KEY,
     user_id BIGINT,
@@ -65,8 +58,7 @@ IF NOT EXISTS warnings
     guild_id BIGINT
 );
 
-CREATE TABLE
-IF NOT EXISTS levels
+CREATE TABLE IF NOT EXISTS levels
 (
     guild_id BIGINT,
     user_id BIGINT,
@@ -74,42 +66,34 @@ IF NOT EXISTS levels
     xp INTEGER DEFAULT 0
 );
 
-CREATE TABLE
-IF NOT EXISTS role_rewards
+CREATE TABLE IF NOT EXISTS role_rewards
 (
     guild_id BIGINT,
     role_id BIGINT,
     required_lvl INTEGER DEFAULT 0
 );
 
-CREATE TABLE
-IF NOT EXISTS birthday
+CREATE TABLE IF NOT EXISTS birthday
 (
     user_id BIGINT,
     birthday TEXT DEFAULT NULL,
     birthday_last_changed BIGINT DEFAULT NULL
 );
 
-CREATE TABLE
-IF NOT EXISTS timezone
+CREATE TABLE IF NOT EXISTS timezone
 (
     user_id BIGINT,
     timezone TEXT DEFAULT NULL,
     timezone_last_changed BIGINT DEFAULT NULL
 );
 
-CREATE TABLE
-IF NOT EXISTS config
+CREATE TABLE IF NOT EXISTS config
 (
     guild_id BIGINT,
     xp_boost INTEGER DEFAULT 1,
     xp_boost_expiry BIGINT DEFAULT 0,
     xp_boost_enabled BOOLEAN DEFAULT TRUE,
     UNIQUE(guild_id)
-
-
 );
 
 DROP TABLE IF EXISTS xp_boosts_user;
-
-
